jam-11 playing with MSR-style "patching"
wonder if I can use global arrays and control them in real-time with the daemon-mode (nice! how fitting!) that would have issues though with the control time I think? the arrays taphy takes are k-time and the daemon mode takes orc code (so we're in instr 0) maybe a helper instrument or something?