yeet the additive, added bass, and reusing vars now
that envelope instrument is cute! and since i use the sequencers in like blocks of code after which i'm done with that signal, i'll just use names and reuse them. no need for numbered variables for each trigger signal. i'd never actually need that. and it's cleaner to just reassign the variable to a new signal than use some signal that's tens of lines away