recursive score instrument
OH! i get it now! the i-pass of an instrument needs to be over
before it can schedule another instrument! so we're passing a delay
value in p2, and the events inside the recursive instances hapen
relative to that time! need to reread order of execution!
(she thought after hours of pure confusion)